requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.1
lxml>=4.9.0
sqlalchemy>=1.4.0
cryptography>=39.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.24.1
aiofiles>=23.1.0
jinja2>=3.1.2
markdown>=3.4.4
python-slugify>=8.0.1 